Fixed colonies and actions not loading after ColonyContext rework
