+++ src/components/v5/common/ActionSidebar/hooks/useGetColonyAction.ts
@@ -28,7 +28,7 @@ export const useGetColonyAction = (transactionHash?: string) => {
   const { refetchTokenBalances } = useUserTokenBalanceContext();
   const isInvalidTx = !isTransactionFormat(transactionHash);
   /* Unfortunately, we need to track polling state ourselves: https://github.com/apollographql/apollo-client/issues/9081#issuecomment-975722271 */
-  const [isPolling, setIsPolling] = useState(isInvalidTx);
+  const [isPolling, setIsPolling] = useState(!isInvalidTx);
 
   const pollInterval = getSafePollingInterval();
 
@@ -49,7 +49,7 @@ export const useGetColonyAction = (transactionHash?: string) => {
   const action = actionData?.getColonyAction;
 
   useEffect(() => {
-    const shouldPool = isInvalidTx && !action;
+    const shouldPool = !isInvalidTx && !action;
 
     setIsPolling(shouldPool);
